backend/requirements-core.txt
# Core dependencies (sufficient for jubu_chat, evaluation, and unit tests)
# For full dependencies including GPU/audio, use: requirements.txt
pydantic[email]==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
numpy>=1.26.0
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
alembic==1.12.0
aiosqlite==0.19.0
loguru>=0.7.0

# Google AI (needed by jubu_chat.chat.models.gemini_model)
google-generativeai>=0.3.0
google-auth>=2.0.0

